Configure the bridge when using non-default thruster configurations.
