+++ vrx_urdf/wamv_gazebo/urdf/thruster_layouts/wamv_gazebo_thruster_config.xacro
@@ -19,6 +19,8 @@
       <max_thrust_cmd>${((x_u + x_uu * max_velocity_mps) * max_velocity_mps)/ 2}</max_thrust_cmd>
       <namespace>${namespace}</namespace>
       <topic>thrusters/${name}/thrust</topic>
+      <!-- Not used by Gazebo but used to run the bridge easily -->
+      <name>${name}</name>
     </plugin>
 
     <plugin
@@ -27,6 +29,8 @@
       <joint_name>${namespace}/${name}_chassis_engine_joint</joint_name>
       <use_velocity_commands>true</use_velocity_commands>
       <topic>${namespace}/thrusters/${name}/pos</topic>
+      <!-- Not used by Gazebo but used to run the bridge easily -->
+      <name>${name}</name>
     </plugin>
 
   </gazebo>
